support LMInterface in pytorch
