+++ espnet/bin/lm_train.py
@@ -59,6 +59,8 @@ def get_parser():
                         help='Filename of validation label data')
     parser.add_argument('--test-label', type=str,
                         help='Filename of test label data')
+    parser.add_argument('--dump-hdf5-path', type=str, default=None,
+                        help='Path to dump a preprocessed dataset as hdf5')
     # LSTMLM training configuration
     parser.add_argument('--opt', default='sgd', type=str,
                         choices=['sgd', 'adam'],
@@ -75,25 +77,44 @@ def get_parser():
                         help="Number of epochs to wait without improvement before stopping the training")
     parser.add_argument('--gradclip', '-c', type=float, default=5,
                         help='Gradient norm threshold to clip')
-    parser.add_argument('--type', type=str, default="lstm", nargs='?', choices=['lstm', 'gru'],
-                        help="Which type of RNN to use")
-    parser.add_argument('--layer', '-l', type=int, default=2,
-                        help='Number of hidden layers')
-    parser.add_argument('--unit', '-u', type=int, default=650,
-                        help='Number of hidden units')
-    parser.add_argument('--dropout-rate', type=float, default=0.5,
-                        help='dropout probability')
     parser.add_argument('--maxlen', type=int, default=40,
                         help='Batch size is reduced if the input sequence > ML')
-    parser.add_argument('--dump-hdf5-path', type=str, default=None,
-                        help='Path to dump a preprocessed dataset as hdf5')
+    # LSTMLM network configuration
+    parser.add_argument('--model-module', type=str, default='legacy',
+                        help='model defined module (default: espnet.nets.xxx_backend.lm.legacy:LegacyRNNLM)')
+    # TODO(karita) remove these options after chainer support
+    # parser.add_argument('--type', type=str, default="lstm", nargs='?', choices=['lstm', 'gru'],
+    #                     help="Which type of RNN to use")
+    # parser.add_argument('--layer', '-l', type=int, default=2,
+    #                     help='Number of hidden layers')
+    # parser.add_argument('--unit', '-u', type=int, default=650,
+    #                     help='Number of hidden units')
+    # parser.add_argument('--dropout-rate', type=float, default=0.5,
+    #                     help='dropout probability')
     return parser
 
 
-def main(args):
-    parser = get_parser()
-    args = parser.parse_args(args)
+# predefined LM dict
+LM_DICT = {
+    "pytorch": {
+        "legacy": "espnet.nets.pytorch_backend.lm.legacy:LegacyRNNLM"
+    },
+    "chainer": {
+        "legacy": None
+    }
+}
 
+
+def main(cmd_args):
+    parser = get_parser()
+    args, _ = parser.parse_known_args(cmd_args)
+
+    # add model-specific arguments dynamically
+    from espnet.utils.dynamic_import import dynamic_import
+    model_class = dynamic_import(args.model_module, LM_DICT[args.backend])
+    model_class.add_arguments(parser)
+
+    args = parser.parse_args(cmd_args)
     # logging info
     if args.verbose > 0:
         logging.basicConfig(
