+++ espnet/nets/pytorch_backend/lm/interface.py
@@ -1,0 +1,3 @@
+import torch
+
+class LMInterface:
